cleanup refactoring diagnostics
