+++ pyproject.toml
@@ -4,6 +4,7 @@ version = "0.2.13"
 
 description = "TraceML: Lightweight runtime bottleneck diagnostics for PyTorch training."
 authors = [{ name = "OptAI UG (haftungsbeschränkt)", email = "support@traceopt.ai" }]
+maintainers = [{ name = "OptAI UG (haftungsbeschränkt)", email = "support@traceopt.ai" }]
 readme = "README.md"
 license = { text = "Apache 2.0" }
 requires-python = ">=3.10"
