+++ src/traceml/renderers/step_time/diagnostics.py
@@ -7,11 +7,14 @@ by both live renderers and summary generation without layering violations.
 
 from traceml.diagnostics.step_time import (
     DEFAULT_THRESHOLDS,
+    LIVE_STEP_TIME_POLICY,
+    SUMMARY_STEP_TIME_POLICY,
     ComputeSignal,
     DiagnosisKind,
     DiagnosisThresholds,
     Severity,
     StepDiagnosis,
+    StepTimeDiagnosisPolicy,
     build_step_diagnosis,
     build_step_diagnosis_result,
     format_cli_diagnosis,
@@ -23,6 +26,9 @@ __all__ = [
     "DiagnosisKind",
     "DiagnosisThresholds",
     "DEFAULT_THRESHOLDS",
+    "LIVE_STEP_TIME_POLICY",
+    "SUMMARY_STEP_TIME_POLICY",
+    "StepTimeDiagnosisPolicy",
     "StepDiagnosis",
     "ComputeSignal",
     "build_step_diagnosis",
