+++ src/traceml/reporting/summaries/step_time_diagnosis.py
@@ -1,6 +1,6 @@
 """Compatibility exports for step-time summary diagnosis helpers."""
 
-from traceml.reporting.sections.step_time.diagnosis import (
+from traceml.diagnostics.step_time.adapters import (
     RankStepSignals,
     build_summary_step_diagnosis,
     build_summary_step_diagnosis_result,
